move mixins into mixins namespace, more cleanup
