+++ lib/trax/model/mixins.rb
@@ -4,9 +4,12 @@ module Trax
       extend ::ActiveSupport::Autoload
 
       autoload :FieldScopes
+      autoload :Freezable
       autoload :IdScopes
+      autoload :Restorable
       autoload :SortByScopes
       autoload :StiEnum
+      autoload :UniqueId
     end
   end
 end
